Extract OrcMetrics utility class
